Refactor Cypress support commands
