+++ cypress/support/commands/payments/index.js
@@ -1,0 +1,3 @@
+export function registerPaymentCommands() {
+  // Reserved for incremental payment command rollout.
+}
