+++ cypress/support/e2e.js
@@ -12,4 +12,3 @@ before(() => {
 after(() => {
   cy.task('removeSetupFile')
 })
-
