+++ cypress/support/helpers/date.js
@@ -1,0 +1,7 @@
+export const getTodayFormatted = () => {
+  const today = new Date();
+  const day = String(today.getDate()).padStart(2, '0');
+  const month = String(today.getMonth() + 1).padStart(2, '0');
+  const year = today.getFullYear();
+  return `${day}-${month}-${year}`;
+};
